Request render before waiting
There was a race condition between the GL thread and the main thread.
If GL thread was much quicker because of scheduling delays, GL thread
already drew the first frame before main thread was reaching resume().

In that case, synch.wait() will just hang because nobody else requested
a frame because the rendering was set to on-demand.

Now, we explicitely request a frame after setting resume=true so we can
make sure a frame gets scheduled with resume=true so we get notified
on the synch.

Change-Id: Id98c9489011d39c6472a98f3e4cb1bfc7b0596ff
Fixes: 30967728